feat: 이모지 클릭후 reload 되던 점 수정
- ServiceHeader 부모 컴포넌트에서 해당하는 로직 실행 및 prop 설정
