+++ src/components/EmojiPicker/EmojiPicker.jsx
@@ -1,28 +1,17 @@
-import React, { useEffect } from 'react';
-import { useParams } from 'react-router-dom';
+import React from 'react';
 import useToggle from '../../hooks/useToggle';
-import usePostData from '../../hooks/usePostData';
-import { postRecipientRollingPaperReactions } from '../../apis/recipientRollingPaperApi';
 import Picker from 'emoji-picker-react';
 import * as S from './EmojiPickerStyle';
-
 import emojiPickerImage from '../../assets/icons/emojiPicker.svg';
 
-const EmojiPicker = () => {
-  const { id } = useParams();
+const EmojiPicker = ({ onEmojiPick }) => {
   const { isOpen, toggleRef, handleToggle, handleClose } = useToggle();
-  const { handlePost, res } = usePostData(postRecipientRollingPaperReactions, [
-    id,
-  ]);
-  const handleEmojiPick = async (emojiObject) => {
+
+  const handleEmojiSelect = (emojiObject, event) => {
+    onEmojiPick(emojiObject);
     handleClose();
-    await handlePost({ emoji: emojiObject.emoji, type: 'increase' });
   };
-  useEffect(() => {
-    if (res) {
-      window.location.reload();
-    }
-  }, [res]);
+
   return (
     <S.EmojiPickerContainer ref={toggleRef}>
       <S.EmojiPickerButton onClick={handleToggle}>
@@ -38,7 +27,7 @@ const EmojiPicker = () => {
         open={isOpen}
         width={300}
         height={400}
-        onEmojiClick={handleEmojiPick}
+        onEmojiClick={handleEmojiSelect}
       />
     </S.EmojiPickerContainer>
   );
